snacks working
