+++ BallsVim/lua/plugins/old/core/ui.lua
@@ -1,13 +1,3 @@
--- Terminal Mappings
-local function term_nav(dir)
-  ---@param self snacks.terminal
-  return function(self)
-    return self:is_floating() and "<c-" .. dir .. ">" or vim.schedule(function()
-      vim.cmd.wincmd(dir)
-    end)
-  end
-end
-
 -- from LazyVim
 ---@param component any
 ---@param text string
@@ -230,121 +220,6 @@ return {
             return icon, hl
           end,
         },
-      })
-    end,
-  },
-  {
-    "snacks.nvim",
-    for_cat = "core.ui",
-    lazy = false,
-    cmd = { "Snacks" },
-    keys = {
-      -- explorer
-      { "<leader>fe", function() Snacks.explorer() end, desc = "Explorer Snacks (cwd)" },
-      { "<leader>e", "<leader>fe", desc = "Explorer Snacks (cwd)", remap = true },
-      -- lazygit: keymaps are in config/keymaps
-      -- [[ Picker ]]
-      { "<leader>,", function() Snacks.picker.buffers() end, desc = "Buffers" },
-      { "<leader>/", function() Snacks.picker.grep() end, desc = "Grep (cwd)" },
-      { "<leader>:", function() Snacks.picker.command_history() end, desc = "Command History" },
-      { "<leader><space>", function() Snacks.picker.files() end, desc = "Find Files (cwd)" },
-      { "<leader>n", function() Snacks.picker.notifications() end, desc = "Notification History" },
-      -- picker find
-      { "<leader>fb", function() Snacks.picker.buffers() end, desc = "Buffers" },
-      { "<leader>fB", function() Snacks.picker.buffers({ hidden = true, nofile = true }) end, desc = "Buffers (all)" },
-      -- { "<leader>fc", LazyVim.pick.config_files(), desc = "Find Config File" },
-      { "<leader>ff", function() Snacks.picker.files() end, desc = "Find Files (cwd)" },
-      { "<leader>fg", function() Snacks.picker.git_files() end, desc = "Find Files (git-files)" },
-      -- { "<leader>fr", LazyVim.pick("oldfiles"), desc = "Recent" },
-      { "<leader>fR", function() Snacks.picker.recent({ filter = { cwd = true }}) end, desc = "Recent (cwd)" },
-      { "<leader>fp", function() Snacks.picker.projects() end, desc = "Projects" },
-      -- picker: git
-      { "<leader>gd", function() Snacks.picker.git_diff() end, desc = "Git Diff (hunks)" },
-      { "<leader>gs", function() Snacks.picker.git_status() end, desc = "Git Status" },
-      { "<leader>gS", function() Snacks.picker.git_stash() end, desc = "Git Stash" },
-      -- picker: grep
-      { "<leader>sb", function() Snacks.picker.lines() end, desc = "Buffer Lines" },
-      { "<leader>sB", function() Snacks.picker.grep_buffers() end, desc = "Grep Open Buffers" },
-      -- { "<leader>sg", LazyVim.pick("live_grep"), desc = "Grep (Root Dir)" },
-      -- { "<leader>sG", LazyVim.pick("live_grep", { root = false }), desc = "Grep (cwd)" },
-      { "<leader>sp", function() Snacks.picker.lazy() end, desc = "Search for Plugin Spec" },
-      -- { "<leader>sw", LazyVim.pick("grep_word"), desc = "Visual selection or word (Root Dir)", mode = { "n", "x" } },
-      -- { "<leader>sW", LazyVim.pick("grep_word", { root = false }), desc = "Visual selection or word (cwd)", mode = { "n", "x" } },
-      -- picker: search
-      { '<leader>s"', function() Snacks.picker.registers() end, desc = "Registers" },
-      { '<leader>s/', function() Snacks.picker.search_history() end, desc = "Search History" },
-      { "<leader>sa", function() Snacks.picker.autocmds() end, desc = "Autocmds" },
-      { "<leader>sc", function() Snacks.picker.command_history() end, desc = "Command History" },
-      { "<leader>sC", function() Snacks.picker.commands() end, desc = "Commands" },
-      { "<leader>sd", function() Snacks.picker.diagnostics() end, desc = "Diagnostics" },
-      { "<leader>sD", function() Snacks.picker.diagnostics_buffer() end, desc = "Buffer Diagnostics" },
-      { "<leader>sh", function() Snacks.picker.help() end, desc = "Help Pages" },
-      { "<leader>sH", function() Snacks.picker.highlights() end, desc = "Highlights" },
-      { "<leader>si", function() Snacks.picker.icons() end, desc = "Icons" },
-      { "<leader>sj", function() Snacks.picker.jumps() end, desc = "Jumps" },
-      { "<leader>sk", function() Snacks.picker.keymaps() end, desc = "Keymaps" },
-      { "<leader>sl", function() Snacks.picker.loclist() end, desc = "Location List" },
-      { "<leader>sM", function() Snacks.picker.man() end, desc = "Man Pages" },
-      { "<leader>sm", function() Snacks.picker.marks() end, desc = "Marks" },
-      { "<leader>sR", function() Snacks.picker.resume() end, desc = "Resume" },
-      { "<leader>sq", function() Snacks.picker.qflist() end, desc = "Quickfix List" },
-      { "<leader>su", function() Snacks.picker.undo() end, desc = "Undotree" },
-      -- picker: ui
-      { "<leader>uC", function() Snacks.picker.colorschemes() end, desc = "Colorschemes" },
-      -- scratch
-      { "<leader>.",  function() Snacks.scratch() end, desc = "Toggle Scratch Buffer" },
-      { "<leader>S",  function() Snacks.scratch.select() end, desc = "Select Scratch Buffer" },
-      { "<leader>dps", function() Snacks.profiler.scratch() end, desc = "Profiler Scratch Buffer" },
-    },
-    after = function(_)
-      require('snacks').setup({
-        bigfile = { enabled = true },
-        dashboard = {
-          preset = {
-            keys = {
-              { icon = " ", key = "f", desc = "Find File", action = ":lua Snacks.dashboard.pick('files')" },
-              { icon = " ", key = "n", desc = "New File", action = ":ene | startinsert" },
-              { icon = " ", key = "g", desc = "Find Text", action = ":lua Snacks.dashboard.pick('live_grep')" },
-              { icon = " ", key = "r", desc = "Recent Files", action = ":lua Snacks.dashboard.pick('oldfiles')" },
-              { icon = " ", key = "s", desc = "Restore Session", section = "session" },
-              { icon = " ", key = "q", desc = "Quit", action = ":qa" },
-            },
-          },
-          sections = {
-            {
-              section = "terminal",
-              cmd = "krabby name golurk --no-title; sleep .8",
-              height = 25,
-              indent = 15,
-            },
-            {
-              section = "keys"
-            }
-          },
-        },
-        explorer = {},
-        indent = { enabled = true },
-        input = { enabled = true },
-        lazygit = {},
-        picker = {},
-        notifier = { enabled = true },
-        quickfile = { enabled = true },
-        scope = { enabled = true },
-        scratch = { enabled = true },
-        scroll = { enabled = true },
-        statuscolumn = { enabled = false },
-        terminal = {
-          win = {
-            keys = {
-              nav_h = { "<C-h>", term_nav("h"), desc = "Go to Left Window", expr = true, mode = "t" },
-              nav_j = { "<C-j>", term_nav("j"), desc = "Go to Lower Window", expr = true, mode = "t" },
-              nav_k = { "<C-k>", term_nav("k"), desc = "Go to Upper Window", expr = true, mode = "t" },
-              nav_l = { "<C-l>", term_nav("l"), desc = "Go to Right Window", expr = true, mode = "t" },
-            },
-          },
-        },
-        toggle = { enabled = true },
-        words = { enabled = true },
       })
     end,
   },
