+++ BallsVim/lua/specs/init.lua
@@ -3,4 +3,5 @@ return {
   { import = "specs.treesitter" },
   { import = "specs.lsp" },
   { import = "specs.ui.which-key" },
+  { import = "specs.ui.snacks" },
 }
