+++ BallsVim/lua/specs/lsp.lua
@@ -37,7 +37,11 @@ return {
     ft = "lua",
     cmd = { "LazyDev" },
     after = function(_)
-      require("lazydev").setup()
+      require("lazydev").setup({
+        library = {
+          { words = { "Snacks" }, path = "snacks.nvim" },
+        },
+      })
     end,
   }
 }
